backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
python-dotenv>=1.0
openai>=1.3
django-cors-headers>=4.0
